+++ dessin_de_courbes/dessin_de_courbes/drawVerticalLine.c
@@ -1,6 +1,6 @@
 #include "headers.h"
 
-void drawVerticalLine(unsigned int xCoordinate, unsigned int yCoordinate, int size, sRGB ** data)
+void drawVerticalLine(unsigned long xCoordinate, unsigned long yCoordinate, int size, sRGB ** data)
 {
 	for (int line = -size; line <= size; line++)
 	{
