+++ dessin_de_courbes/dessin_de_courbes/headers.h
@@ -10,9 +10,9 @@ void bitmapFreeData(sRGB **Data);
 //	Drawing Functions
 void drawMargin(unsigned int margin, sRGB ** data, unsigned int width, unsigned int height);
 void drawAxe(double minX, double maxX, double stepX, double minY, double maxY, double stepY, sRGB ** data, unsigned long width, unsigned long height, unsigned int margin);
-void drawHorizontalLine(unsigned int xCoordinate, unsigned int yCoordinate, unsigned int size, sRGB ** data);
-void drawVerticalLine(unsigned int xCoordinate, unsigned int yCoordinate, int size, sRGB ** data);
-void drawPoint(unsigned int xCoordinate, unsigned int yCoordinate, int size, sRGB ** data, sRGB color);
+void drawHorizontalLine(unsigned long xCoordinate, unsigned long yCoordinate, int size, sRGB ** data);
+void drawVerticalLine(unsigned long xCoordinate, unsigned long yCoordinate, int size, sRGB ** data);
+void drawPoint(unsigned long xCoordinate, unsigned long yCoordinate, int size, sRGB ** data, sRGB color);
 
 //	File Functions
 void fileWrite(unsigned long width, unsigned long height, sRGB ** data);
